Porting changes from #240 and other fixes. All exports working
